+++ graphene_django_crud/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from .types import DjangoGrapheneCRUD, resolver_hints
+from .types import DjangoGrapheneCRUD, DjangoCRUDObjectType, resolver_hints
 from .utils import where_input_to_Q, order_by_input_to_args
 from .base_types import DefaultConnection, EmptyDefaultConnection
 
@@ -7,6 +7,7 @@ __version__ = "1.2.1"
 
 __all__ = (
     "__version__",
+    "DjangoCRUDObjectType",
     "DjangoGrapheneCRUD",
     "resolver_hints",
     "where_input_to_Q",