+++ tests/djangocontribauth/schema.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 from django.contrib.auth.models import User, Group
-from graphene_django_crud import DjangoGrapheneCRUD, where_input_to_Q, resolver_hints
+from graphene_django_crud import DjangoCRUDObjectType, where_input_to_Q, resolver_hints
 import graphene
 
 
-class UserType(DjangoGrapheneCRUD):
+class UserType(DjangoCRUDObjectType):
     class Meta:
         model = User
 
@@ -22,7 +22,7 @@ class UserType(DjangoGrapheneCRUD):
         return super().mutate(parent, info, instance, data, *args, **kwargs)
 
 
-class GroupType(DjangoGrapheneCRUD):
+class GroupType(DjangoCRUDObjectType):
     class Meta:
         model = Group