+++ CHANGELOG.md
@@ -9,10 +9,13 @@ All notable changes to this project will be documented in this file.
 - Relay integration
 - Possibility to extend connectionType
 - Possibility of having a list field instead of a connection field
-- Replace the methods before_xxx, after_xxx methods to the methods mutate, create, update, delete 
+- Replace the methods before_xxx, after_xxx methods to the methods mutate, create,
+  update, delete
 
 ### Changed
 
+- DjangoGrapheneCRUD class has been renamed to DjangoCRUDObjectType, so the name
+  "DjangoGrapheneCRUD" is deprecated.
 - Graphql type <typeName>NodeType is renamed to <typeName>Connection
 - Graphql type <typeName>MutationType is renamed to <typeName>Mutation
 - The methods before_xxx, after_xxx are deprecated